requirements.txt
wget
requests
Pillow